requirements.txt
diffusers==0.21.4
transformers==4.33.2
torch==2.0.1
accelerate==0.24.1
safetensors==0.4.0
huggingface-hub==0.17.3
numpy<2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
reportlab==4.1.0
stability-sdk==0.8.5
openai>=1.3.0
requests>=2.31.0
